refactoring createFolders on start server
