+++ bin/server.js
@@ -1,21 +1,8 @@
 const app = require('../app')
 const { connectMongo } = require('../db/connection')
+const { createFolderIsNotExist } = require('../helpers/createFolders')
 
-const fs = require('fs').promises
-const { UPLOAD_DIR, AVATARS_DIR } = require('../routes/api/upload')
-
-const isAccessible = (path) => {
-  return fs
-    .access(path)
-    .then(() => true)
-    .catch(() => false)
-}
-
-const createFolderIsNotExist = async (folder) => {
-  if (!(await isAccessible(folder))) {
-    await fs.mkdir(folder)
-  }
-}
+const { UPLOAD_DIR, AVATARS_DIR } = require('../helpers/uploads')
 
 const PORT = process.env.PORT || 3000
 const start = async () => {
